+++ EventListener/ValidateSchemaListener.php
@@ -11,26 +11,6 @@ class ValidateSchemaListener
     /** @var ValidateSchemaService */
     protected $validateSchema;
 
-    /** @var array */
-    protected $excludedEntities = array();
-
-    /** @var array */
-    protected $excludedProperties = array();
-
-    /**
-     * Messages from Doctrine\ORM\Tools\SchemaValidator
-     *
-     * @var array
-     */
-    protected $schemaValidatorMessages = array(
-        'The field \'%s\' ',
-        'The field %s ',
-        'The association %s ',
-        'Cannot map association \'%s\' ',
-        'The mappings %s and ',
-        'If association %s '
-    );
-
     /**
      * @param ValidateSchemaService $validateSchema
      */